test(ios): share source slicing helper
