+++ Weird Parts IOS/Weird Parts IOSTests/WarehouseWizardProgressAccessibilityTests.swift
@@ -285,30 +285,8 @@ final class WarehouseWizardProgressAccessibilityTests: XCTestCase {
         return String(afterStart[..<end])
     }
 
-    /// Extracts the brace-balanced body that follows the first occurrence of
-    /// `anchor`, so assertions stay scoped to the code under test.
     private static func braceBalancedBody(after anchor: String, in source: String) throws -> String {
-        guard let anchorRange = source.range(of: anchor) else {
-            throw XCTSkip("Expected anchor \(anchor) in source")
-        }
-        guard let openBrace = source[anchorRange.upperBound...].firstIndex(of: "{") else {
-            throw XCTSkip("Expected opening brace after \(anchor)")
-        }
-
-        var depth = 0
-        var index = openBrace
-        while index < source.endIndex {
-            let char = source[index]
-            if char == "{" { depth += 1 }
-            if char == "}" { depth -= 1 }
-            let next = source.index(after: index)
-            if depth == 0 {
-                return String(source[openBrace..<next])
-            }
-            index = next
-        }
-
-        throw XCTSkip("Expected closing brace for \(anchor)")
+        try TestSourceSlicer.braceBalancedBody(after: anchor, in: source)
     }
 
     private static func readWarehouseSource(
